Use remote Docker server for Travis builds
This is likely to be a lot more stable since we won't be bumping up against
resource limits by trying to run Docker in UML on the Travis worker.

Note that this does not work for forks. For any pull requests from forks,
we still use the old UML method.